Create MoveDuplicateFiles
Using MD5 hash comparison, this will recursively scan folders and subfolders for duplicate files. When it finds a duplicate, it will then move it to the folder you specified on line 5. It will then further create its' own subfolder depending on the extension, to keep the identified filetypes separated. All folders required for this script will be created if they don't already exist.

Line 2 = Set the file directory you want to scan
Line 5 = Set the directory that will contain the duplicate files to be moved